Platforms Update (18)
This level is now over 6 months long… but still not finished.

FIXED :
-I did a lot of code changes, so not much was fixed.

BROKE :
-The game will not run… I suppose that’s considered breaking?

ADDED :
-More platforms
-A goal thing that you should be able to bump into.  Pretty sure that’s
the thing that’s causing an error.  There was also a ladder class which
I deleted before this, with the same class as Steps and Goals.

I need to also go on Piskel and make some textures for the things and
the stuff.

How long have I been up;
-Porpoise